requirements.txt
flask
Flask-Cors
python-dotenv
torch
langchain
langchain-community
langchain-huggingface
langchainhub
huggingface
pypdf
sentence-transformers
faiss-cpu
llama-index
llama-index-llms-huggingface
llama-index-embeddings-huggingface
llama-index-vector-stores-faiss
werkzeug